+++ src/sc_crawler/inspector.py
@@ -20,7 +20,7 @@ from yaml import safe_load as yaml_safe_load
 from .inspector_helpers import _get_cpu_cache_info
 from .logger import logger
 from .table_bases import ServerBase
-from .table_fields import DdrGeneration, Disk, StorageType
+from .table_fields import DdrGeneration, Disk, Parallelism, StorageType
 
 if TYPE_CHECKING:
     from .tables import Server
@@ -260,7 +260,9 @@ def inspect_server_benchmarks(server: "Server") -> List[dict]:
                     config = {
                         "algo": algo,
                         "compression_level": None if level == "null" else int(level),
-                        "multithread": True if threads > 1 else False,
+                        "cores": Parallelism.MULTI
+                        if threads > 1
+                        else Parallelism.SINGLE,
                     }
                     if data.get("extra_args", {}).get("block_size"):
                         config["block_size"] = data["extra_args"]["block_size"]
@@ -285,9 +287,17 @@ def inspect_server_benchmarks(server: "Server") -> List[dict]:
             scores = json.load(fp)
         geekbench_version = _server_framework_meta(server, framework)["version"]
         for cores, workloads in scores.items():
+            parallelism = (
+                Parallelism.SINGLE
+                if cores == "Single-Core Performance"
+                else Parallelism.MULTI
+            )
             for workload, values in workloads.items():
                 workload_fields = {
-                    "config": {"cores": cores, "framework_version": geekbench_version},
+                    "config": {
+                        "cores": parallelism,
+                        "framework_version": geekbench_version,
+                    },
                     "score": float(values["score"]),
                 }
                 if values.get("description"):